FM-366: Split into multiple modules
